Split part of concepts.h into range_concepts.h
Due to <ranges> this header is quite expensive to compile, and the
ranges parts aren't needed by many users of concepts.h - and in fact
isn't needed at all in the public API outside of the (deprecated for
public use) mem_ops.h

Move the parts of concepts.h related to strong types to that header.

Remove the opportunistic exposure of checked accessors in strong type;
it wasn't used outside of tests and the concept is somewhat complicated.
