+++ src/lib/pubkey/ec_group/ec_scalar.h
@@ -12,6 +12,7 @@
 #include <memory>
 #include <optional>
 #include <span>
+#include <string_view>
 #include <vector>
 
 namespace Botan {
