+++ src/lib/utils/concepts.h
@@ -11,24 +11,8 @@
 
 #include <botan/types.h>
 #include <concepts>
-#include <iosfwd>
-#include <ranges>
-#include <span>
-#include <type_traits>
 
 namespace Botan {
-
-template <typename T, typename Tag, typename... Capabilities>
-class Strong;
-
-template <typename... Ts>
-struct is_strong_type : std::false_type {};
-
-template <typename... Ts>
-struct is_strong_type<Strong<Ts...>> : std::true_type {};
-
-template <typename... Ts>
-constexpr bool is_strong_type_v = is_strong_type<std::remove_const_t<Ts>...>::value;
 
 template <typename T0 = void, typename... Ts>
 struct all_same {
@@ -49,106 +33,6 @@ struct AutoDetect {
 };
 
 }  // namespace detail
-
-namespace ranges {
-
-/**
- * Models a std::ranges::contiguous_range that (optionally) restricts its
- * value_type to ValueT. In other words: a stretch of contiguous memory of
- * a certain type (optional ValueT).
- */
-template <typename T, typename ValueT = std::ranges::range_value_t<T>>
-concept contiguous_range = std::ranges::contiguous_range<T> && std::same_as<ValueT, std::ranges::range_value_t<T>>;
-
-/**
- * Models a std::ranges::contiguous_range that satisfies
- * std::ranges::output_range with an arbitrary value_type. In other words: a
- * stretch of contiguous memory of a certain type (optional ValueT) that can be
- * written to.
- */
-template <typename T, typename ValueT = std::ranges::range_value_t<T>>
-concept contiguous_output_range = contiguous_range<T, ValueT> && std::ranges::output_range<T, ValueT>;
-
-/**
- * Models a range that can be turned into a std::span<>. Typically, this is some
- * form of ranges::contiguous_range.
- */
-template <typename T>
-concept spanable_range = std::constructible_from<std::span<const std::ranges::range_value_t<T>>, T>;
-
-/**
- * Models a range that can be turned into a std::span<> with a static extent.
- * Typically, this is a std::array or a std::span derived from an array.
- */
-// clang-format off
-template <typename T>
-concept statically_spanable_range = spanable_range<T> &&
-                                    decltype(std::span{std::declval<T&>()})::extent != std::dynamic_extent;
-
-// clang-format on
-
-/**
- * Find the length in bytes of a given contiguous range @p r.
- */
-inline constexpr size_t size_bytes(const spanable_range auto& r) {
-   return std::span{r}.size_bytes();
-}
-
-/**
-* Throws an exception indicating that the attempted read or write was invalid
-*/
-[[noreturn]] void BOTAN_UNSTABLE_API memory_region_size_violation();
-
-/**
- * Check that a given range @p r has a certain statically-known byte length. If
- * the range's extent is known at compile time, this is a static check,
- * otherwise a runtime argument check will be added.
- *
- * @throws Invalid_Argument  if range @p r has a dynamic extent and does not
- *                           feature the expected byte length.
- */
-template <size_t expected, spanable_range R>
-inline constexpr void assert_exact_byte_length(const R& r) {
-   const std::span s{r};
-   if constexpr(statically_spanable_range<R>) {
-      static_assert(s.size_bytes() == expected, "memory region does not have expected byte lengths");
-   } else {
-      if(s.size_bytes() != expected) {
-         memory_region_size_violation();
-      }
-   }
-}
-
-/**
- * Check that a list of ranges (in @p r0 and @p rs) all have the same byte
- * lengths. If the first range's extent is known at compile time, this will be a
- * static check for all other ranges whose extents are known at compile time,
- * otherwise a runtime argument check will be added.
- *
- * @throws Invalid_Argument  if any range has a dynamic extent and not all
- *                           ranges feature the same byte length.
- */
-template <spanable_range R0, spanable_range... Rs>
-inline constexpr void assert_equal_byte_lengths(const R0& r0, const Rs&... rs)
-   requires(sizeof...(Rs) > 0)
-{
-   const std::span s0{r0};
-
-   if constexpr(statically_spanable_range<R0>) {
-      constexpr size_t expected_size = s0.size_bytes();
-      (assert_exact_byte_length<expected_size>(rs), ...);
-   } else {
-      const size_t expected_size = s0.size_bytes();
-      const bool correct_size =
-         ((std::span<const std::ranges::range_value_t<Rs>>{rs}.size_bytes() == expected_size) && ...);
-
-      if(!correct_size) {
-         memory_region_size_violation();
-      }
-   }
-}
-
-}  // namespace ranges
 
 namespace concepts {
 
@@ -184,19 +68,6 @@ concept has_empty = requires(T a) {
    { a.empty() } -> std::same_as<bool>;
 };
 
-// clang-format off
-template <typename T>
-concept has_bounds_checked_accessors = container<T> && (
-                                          requires(T a, const T ac, typename T::size_type s) {
-                                             { a.at(s) } -> std::same_as<typename T::value_type&>;
-                                             { ac.at(s) } -> std::same_as<const typename T::value_type&>;
-                                          } ||
-                                          requires(T a, const T ac, typename T::key_type k) {
-                                             { a.at(k) } -> std::same_as<typename T::mapped_type&>;
-                                             { ac.at(k) } -> std::same_as<const typename T::mapped_type&>;
-                                          });
-// clang-format on
-
 template <typename T>
 concept resizable_container = container<T> && requires(T& c, typename T::size_type s) {
    T(s);
@@ -209,24 +80,6 @@ concept reservable_container = container<T> && requires(T& c, typename T::size_t
 template <typename T>
 concept resizable_byte_buffer =
    contiguous_container<T> && resizable_container<T> && std::same_as<typename T::value_type, uint8_t>;
-
-template <typename T>
-concept streamable = requires(std::ostream& os, T a) { os << a; };
-
-template <class T>
-concept strong_type = is_strong_type_v<T>;
-
-template <class T>
-concept contiguous_strong_type = strong_type<T> && contiguous_container<T>;
-
-template <class T>
-concept integral_strong_type = strong_type<T> && std::integral<typename T::wrapped_type>;
-
-template <class T>
-concept unsigned_integral_strong_type = strong_type<T> && std::unsigned_integral<typename T::wrapped_type>;
-
-template <typename T, typename Capability>
-concept strong_type_with_capability = T::template has_capability<Capability>();
 
 }  // namespace concepts
