+++ src/lib/utils/mem_ops.h
@@ -9,7 +9,7 @@
 #define BOTAN_MEMORY_OPS_H_
 
 #include <botan/assert.h>
-#include <botan/concepts.h>
+#include <botan/range_concepts.h>
 #include <botan/types.h>
 #include <array>
 #include <cstring>
